move or_empty deserializer to ruma-common, add tests
